third_party/WebKit/LayoutTests/platform/win/storage/quota/storagemanager-estimate-expected.txt

This is a testharness.js-based test.
PASS estimate() method exists and returns a Promise 
PASS estimate() resolves to dictionary with members 
PASS estimate() shows usage increase after large value is stored 
Harness: the test ran to completion.
